feat: calculating control key
