+++ exercise/kotlin/day17/src/test/kotlin/eid/EIDTest.kt
@@ -8,6 +8,7 @@ import io.kotest.matchers.types.shouldBeInstanceOf
 import net.jqwik.api.*
 import net.jqwik.api.constraints.IntRange
 import net.jqwik.api.constraints.StringLength
+import org.junit.jupiter.api.Test
 
 class EIDTest {
 
@@ -109,6 +110,11 @@ class EIDTest {
         EID.parse(input).shouldBeLeft().shouldBeInstanceOf<ParsingError.InvalidSerialNumber>()
     }
 
+    @Test
+    fun calculatesControlKeyForEidCandidate() {
+        EID.calculateControlKey("198007xx").shouldBeRight() shouldBe 67
+    }
+
     @Provide
     fun validEidBuilders(): Arbitrary<EidStringBuilder> {
         return Combinators.combine(validGenders(), validYears(), validSerialNumbers())
@@ -157,68 +163,3 @@ class EIDTest {
     val digitCharacters: CharRange = '0'..'9'
 }
 
-class EidStringBuilder {
-    private var gender: String = ""
-    private var year: String = ""
-    private var serialNumber: String = ""
-    private var controlKeyOverride: String? = null
-
-    fun withGender(gender: ElfGender): EidStringBuilder {
-        this.gender = toGenderString(gender)
-        return this
-    }
-
-    fun withGenderString(genderString: String): EidStringBuilder {
-        this.gender = genderString
-        return this
-    }
-
-    fun withYear(year: Int): EidStringBuilder {
-        this.year = toYearString(year)
-        return this
-    }
-
-    fun withYearString(yearString: String): EidStringBuilder {
-        this.year = yearString
-        return this
-    }
-
-    fun withSerialNumber(serialNumber: Int): EidStringBuilder {
-        this.serialNumber = toSerialNumberString(serialNumber)
-        return this
-    }
-
-    fun withSerialNumberString(serialNumberString: String): EidStringBuilder {
-        this.serialNumber = serialNumberString
-        return this
-    }
-
-    fun withControlKeyOverride(controlKeyOverride: String): EidStringBuilder {
-        this.controlKeyOverride = controlKeyOverride
-        return this
-    }
-
-    fun build(): String {
-        var controlKey = "00" //TODO calculate
-        if (controlKeyOverride != null) {
-            controlKey = controlKeyOverride!!
-        }
-        return "$gender$year$serialNumber$controlKey"
-    }
-
-    private fun toGenderString(elfGender: ElfGender): String {
-        return when (elfGender) {
-            ElfGender.Sloubi -> "1"
-            ElfGender.Gagna -> "2"
-            ElfGender.Catact -> "3"
-        }
-    }
-
-    private fun toYearString(year: Int): String {
-        return year.toString().padStart(2, '0')
-    }
-
-    private fun toSerialNumberString(serialNumber: Int): String {
-        return serialNumber.toString().padStart(3, '0')
-    }
-}
